I'm really frustrated so I've got to save this and move on
Basically I had pages for scan -> update details -> scan again or if the
book already existed then scan -> book details, but whenever I did scan
again, the scan page would get redirected to, but it wouldn't show anything
it's like the React componets didn't mount, but the mounting code was being
called so maybe the dom wasn't loaded when the event went. It was really
confusing and I don't know if it's a react, rails, rails dev, or me problem
so I'm just moving on and trying to start over I guess. So now the scan
page doesn't redirect correctly at all, but at least the fixtures work.

The same problem is going on still in the static pages controller when
redirecting to /scan/:id, but I have no idea what the deal is and I don't
have patience for it right now.